+++ crates/wdl-lint/src/rules/shellcheck.rs
@@ -18,6 +18,8 @@ use serde_json;
 use tracing::debug;
 use wdl_analysis::Diagnostics;
 use wdl_analysis::Document;
+use wdl_analysis::Example;
+use wdl_analysis::LabeledSnippet;
 use wdl_analysis::VisitReason;
 use wdl_analysis::Visitor;
 use wdl_analysis::diagnostics::unknown_type;
@@ -236,35 +238,33 @@ impl Rule for ShellCheckRule {
          Following its recommendations will increase the robustness of your command sections."
     }
 
-    fn examples(&self) -> &'static [&'static str] {
-        &[
-            r#"```wdl
-version 1.2
+    fn examples(&self) -> &'static [Example] {
+        &[Example {
+            negative: LabeledSnippet {
+                label: None,
+                snippet: r#"version 1.2
 
 task say_hello {
-    meta {}
-
     # Triggers SC2154
     command <<<
         echo "Hello $name"
     >>>
 }
-```"#,
-            r#"Use instead:
-
-```wdl
-version 1.2
+"#,
+            },
+            revised: Some(LabeledSnippet {
+                label: None,
+                snippet: r#"version 1.2
 
 task say_hello {
-    meta {}
-
     command <<<
         name=World
         echo "Hello $name"
     >>>
 }
-```"#,
-        ]
+"#,
+            }),
+        }]
     }
 
     fn tags(&self) -> TagSet {